begin separating out concrete types in metadata
The end goal of this patch series is to have a clean way of representing
data models for "concrete" types -- i.e. models that have been validated
and any relations expanded into real objects. For example, a user passes
a byte array containing a YAML document describing an object. That YAML
document is unmarshalled into an api/types/Object struct. It contains
simple identifier for partition, project, object type, etc. When we pass
create/update instructions to backend storage, we want to have exploded
that api/types/Object struct into a struct containing valid pointers to
Partition and ObjectType models so we don't continually need to re-fetch
partition and object type models.
