+++ pkg/metadata/types/README.md
@@ -1,0 +1,2 @@
+The `pkg/metadata/types` package contains concrete structs used internally to
+the `runm-metadata` server and its storage layer.
